Updated SM to have its own get/set interrupt functions.
Modified the task swap routine to save the SDA, and then set the SM SDA to be the SDA SM had upon launching the first task.

Modified the task swap routine to use these new get/set interrupt functions.

Now we have ^C support within SM, it resets the screen.

Init checks that STDIO are not redirected and fails if they are.

Added an explaination for as to why we set i22h to a temporary pointer before we spawn the sessions (if not COMMAND.COM in the future, and it exits, will return to that handler which eventually will respawn that task).

Only before switching to session 1 in init, SM now sets itself to be its own parent and modifies its int 22, 23 and 24 addresses in both the PSP and PSDA.  Thus, if we terminate before launching tasks at any point, all resources will be properly cleaned (and the installed Int 22h will not affect this as this only affects child tasks).